Major changes: continuing to add histogram, percentile and stretching

+++ defsui.h
@@ -125,15 +125,15 @@ struct WidgetsStates
 
 struct ImageParams
 {
-    uint32_t width;
-    uint32_t height;
-    double   bzero;
-    double   bscale;
-    size_t   HDUBaseOffset;
-    size_t   maxDataBufferSize;
-    uint32_t hduIndex;
-    int8_t   bitpix;
-    bool     grayscale;
+    uint32_t        width;
+    uint32_t        height;
+    long double     bzero;
+    long double     bscale;
+    size_t          HDUBaseOffset;
+    size_t          maxDataBufferSize;
+    uint32_t        hduIndex;
+    int8_t          bitpix;
+    bool            grayscale;
 };
 
 #endif // DEFSUI_H
